changed cmp sources, added url-under-cursor feat
kitty is no longer used as a cmp source as it was slowing down telescope
windows. would be lovely to be able to support it in the future.

english dictionary added as a cmp source but disabled for now - need to
find a way to toggle it on and off as it takes up a huge amount of the
completion popup.

also finally added the ability to navigate to urls with "gx"